+++ alh_pdf_view_android/android/src/main/kotlin/com/example/alh_pdf_view/AlhPdfView.kt
@@ -109,6 +109,10 @@ internal class AlhPdfView(
             .pageSnap(alhPdfViewConfiguration.pageSnap)
             .enableDoubletap(alhPdfViewConfiguration.enableDoubleTap)
             .defaultPage(defaultPage)
+            .onTap {
+                alhPdfViewChannel.invokeMethod("onTap", null)
+                true
+            }
             .onPageChange { page, total ->
                 if (hasSetPageWithAnimation) {
                     hasSetPageWithAnimation = false