+++ alh_pdf_view_ios/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 2.1.0
+* Added `onTap` as callback for iOS when tapping pdf suggested by https://github.com/alhappler/alh_pdf_view/pull/36.
+
 ## 2.0.0
 
 * Added iOS implementation, see more on changelogs in [alh_pdf_view](https://pub.dev/packages/alh_pdf_view/changelog)