+++ alh_pdf_view_platform_interface/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 2.1.0
+* Added `onTap` as callback for Android and iOS when tapping pdf suggested by https://github.com/alhappler/alh_pdf_view/pull/36.
+
 ## 2.0.0
 
 * Added platform interface, see more on changelogs in [alh_pdf_view](https://pub.dev/packages/alh_pdf_view/changelog)